+++ BTCPayServer/Monero/Payments/MoneroSupportedPaymentMethod.cs
@@ -4,7 +4,7 @@ namespace BTCPayServer.Payments.Monero
     {
 
         public string CryptoCode { get; set; }
-        public int AccountIndex { get; set; }
+        public long AccountIndex { get; set; }
         public PaymentMethodId PaymentId => new PaymentMethodId(CryptoCode, MoneroPaymentType.Instance);
     }
 }
